+++ packages/stream_chat_flutter/lib/src/message_input/clear_input_item_button.dart
@@ -31,6 +31,7 @@ class ClearInputItemButton extends StatelessWidget {
         hoverElevation: 0,
         onPressed: onTap,
         fillColor:
+            // ignore: deprecated_member_use
             _streamChatTheme.colorTheme.textHighEmphasis.withOpacity(0.5),
         child: Center(
           child: StreamSvgIcon.close(