+++ packages/stream_chat_flutter/test/src/message_input/attachment_button_test.dart
@@ -1,6 +1,6 @@
+import 'package:alchemist/alchemist.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:golden_toolkit/golden_toolkit.dart';
 import 'package:stream_chat_flutter/src/message_input/attachment_button.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
@@ -33,22 +33,21 @@ void main() {
     expect(count, 1);
   });
 
-  testGoldens('golden test for AttachmentButton', (tester) async {
-    await tester.pumpWidget(
-      MaterialAppWrapper(
-        home: Scaffold(
-          body: Center(
-            child: AttachmentButton(
-              color: StreamChatThemeData.light()
-                  .messageInputTheme
-                  .actionButtonIdleColor,
-              onPressed: () {},
-            ),
+  goldenTest(
+    'golden test for AttachmentButton',
+    fileName: 'attachment_button_0',
+    constraints: const BoxConstraints.tightFor(width: 50, height: 50),
+    builder: () => MaterialAppWrapper(
+      home: Scaffold(
+        body: Center(
+          child: AttachmentButton(
+            color: StreamChatThemeData.light()
+                .messageInputTheme
+                .actionButtonIdleColor,
+            onPressed: () {},
           ),
         ),
       ),
-    );
-
-    await screenMatchesGolden(tester, 'attachment_button_0');
-  });
+    ),
+  );
 }