+++ packages/stream_chat_flutter_core/CHANGELOG.md
@@ -5,6 +5,10 @@
 - Added `StreamPollController` to create and manage a poll based on the passed configs.
 - Added `StreamPollVoteListController` to manage the list of votes for a poll.
 
+🔄 Changed
+
+- Updated minimum Flutter version to 3.24.5 for the SDK.
+
 ## 8.3.0
 
 - Updated `stream_chat` dependency to [`8.3.0`](https://pub.dev/packages/stream_chat/changelog).